% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cn_stopwords}
\alias{cn_stopwords}
\title{1893 Chinese stopwords}
\format{a string vector}
\source{
\url{https://blog.csdn.net/shijiebei2009/article/details/39696571"}
}
\usage{
cn_stopwords
}
\description{
1893 Chinese stopwords
}
\keyword{datasets}
