% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnwordcloud.R
\name{cnwordcloud}
\alias{cnwordcloud}
\title{This function allows you to generate a wordcloud in Chinese.}
\usage{
cnwordcloud(text, stopwords = NULL, nwords = 50)
}
\arguments{
\item{text}{a vector of strings.}

\item{stopwords}{a dictionary of stopwords, default to cn_stopwords.}

\item{nwords}{the number of words to include in the wordcloud, default to 50.}
}
\value{
none.
}
\description{
This function allows you to generate a wordcloud in Chinese.
}
\examples{
library(cnwordcloud)
data(sample_text)
cnwordcloud(sample_text)
}
\keyword{Chinese.}
\keyword{wordcloud,}
