% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_text}
\alias{sample_text}
\title{A sample text, Xi's speech at 2019 CDAC}
\format{a string vector}
\source{
\url{https://news.sina.com.cn/gov/xlxw/2019-05-15/doc-ihvhiqax8868898.shtml}
}
\usage{
sample_text
}
\description{
A sample text, Xi's speech at 2019 CDAC
}
\keyword{datasets}
